+++ crates/vmi-os-windows/src/offsets/mod.rs
@@ -448,6 +448,8 @@ offsets! {
 
         struct _EPROCESS {
             UniqueProcessId: Field,
+            CreateTime: Field,              // _LARGE_INTEGER
+            ExitTime: Field,                // _LARGE_INTEGER
             ActiveProcessLinks: Field,      // _LIST_ENTRY
             SessionProcessLinks: Field,     // _LIST_ENTRY
             SectionBaseAddress: Field,